Clarified keyboard extensions document
This program could be way easier to use than it currently is. There's a lot of friction with having to go back and click the model you want to add then clicking the screen. I'm surprised it's not an option to ctrl-click to add another of the previous model again, or for there to be another shortcut to add another outgoing process. 

And then I'm surprised that the location of where to put the keymap.json wasn't intuitive and proven to immediately work. I read the original documentation as plopping a keymap.json in the configuration path. This did not work, so after digging, I learned about creating shortcuts with extensions. That worked. 

So I'm updating the with a tested, clear and working process for creating a keyboard shortcut. 

If there's a better way please correct me. I just wanted a keyboard shortcut to add an outgoing process (with flowcharts).